MvcSocialWeb: Добавлено перечисление Months

+++ MvcSocialWeb/MapperProfile.cs
@@ -14,7 +14,7 @@ namespace MvcSocialWeb
             CreateMap<RegisterViewModel, User>()
                 .ForMember(user => user.Email, opt => opt.MapFrom(model => model.EmailReg))
                 .ForMember(user => user.UserName, opt => opt.MapFrom(model => model.Login))
-                .ForMember(user => user.BirthDate, opt => opt.MapFrom(model => new DateTime(model.Year, model.Month, model.Date)))
+                .ForMember(user => user.BirthDate, opt => opt.MapFrom(model => new DateTime((int)model.Year, (int)model.Month, (int)model.Date)))
                 .ForMember(user => user.PasswordHash, opt => opt.MapFrom(model => model.PasswordReg.GetHashCode()));
         }
     }
